Push sort & paginate down into service
