+++ src/api/Services/ICourseSearchService.cs
@@ -1,11 +1,12 @@
-﻿using System.Linq;
-using GovUk.Education.SearchAndCompare.Domain.Filters;
+﻿using GovUk.Education.SearchAndCompare.Domain.Filters;
+using GovUk.Education.SearchAndCompare.Domain.Lists;
 using GovUk.Education.SearchAndCompare.Domain.Models;
 
 namespace GovUk.Education.SearchAndCompare.Api.Services
 {
     public interface ICourseSearchService
     {
-        IQueryable<Course> GetFilteredCourses(QueryFilter filter);
+        PaginatedList<Course> GetFilteredCourses(QueryFilter filter);
+        int GetFilteredCourseCount(QueryFilter filter);
     }
 }
